started implementing pso
